# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Demo Loans System",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api/{version}",
            "description": "Api for loans system",
            "variables": {
                "version": {
                    "enum": [
                        "v1"
                    ],
                    "default": "v1"
                }
            }
        }
    ],
    "paths": {
...
        "/payment/list": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Get list payments",
                "operationId": "5d6780b6c4a330192be50a9b4ba37293",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Request Header",
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    },
                    {
                        "name": "items",
                        "in": "query",
                        "description": "Paginate with items.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of payments"
                    },
                    "401": {
                        "description": "Unauthorize."
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/payment/create": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create payment for Loan",
                "description": "Create payment for Loan",
                "operationId": "ae56ba49fd7adc3275301a3a1b298f3c",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Request Header",
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/payment"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "402": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/payment/detail/{paymentId}": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Get Payment Detail",
                "operationId": "e2ec66be6776b3e2983ad0ed1bed4665",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Request Header",
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    },
                    {
                        "name": "paymentId",
                        "in": "path",
                        "description": "Payment Id.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of service provider"
                    },
                    "401": {
                        "description": "Unauthorize."
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/payment/approve": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Approve payment",
                "description": "Approve payment",
                "operationId": "61a6c3206cee8df724790dc8ff74ad02",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Request Header",
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/payment_approve"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "402": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "payment": {
                "title": "Payment Model",
                "properties": {
                    "loan_id": {
                        "schema": "Payment",
                        "description": "loan_id",
                        "type": "integer",
                        "example": "1"
                    },
                    "amount": {
                        "schema": "Payment",
                        "description": "amount",
                        "type": "float",
                        "example": "10.00"
                    },
                    "payment_mode": {
                        "schema": "Payment",
                        "description": "payment_mode",
                        "type": "string",
                        "example": "paid"
                    },
                    "description": {
                        "schema": "Payment",
                        "description": "description",
                        "type": "string",
                        "example": "description"
                    }
                },
                "type": "object"
            },
            "payment_approve": {
                "title": "Payment approve Model",
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "description": "payment id",
                                "type": "integer",
                                "example": "1"
                            },
                            "loan_id": {
                                "description": "loan id",
                                "type": "integer",
                                "example": "1"
                            },
                            "status": {
                                "description": "status",
                                "type": "integer",
                                "default": 1,
                                "enum": [
                                    0,
                                    1
                                ]
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "description": "Header value: `Bearer {access_token}`",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}